Move fetcher with indexer stuff and his friend classes
